+++ src/Validators/ConfigurationValidatorError.ts
@@ -1,0 +1,2 @@
+export class ConfigurationValidatorError extends Error {
+}