+++ src/index.ts
@@ -20,7 +20,9 @@ export * from "./Sources/SourcesAggregator";
 
 export * from "./Validators/ArrayValidator";
 export * from "./Validators/BooleanValidator";
+export * from "./Validators/ConfigurationValidatorError";
 export * from "./Validators/ConfigurationValidatorInterface";
+export * from "./Validators/EnumerableValidator";
 export * from "./Validators/NumberValidator";
 export * from "./Validators/RegularExpressionValidator";
 export * from "./Validators/StringValidator";