+++ tests/index.test.ts
@@ -17,6 +17,8 @@ test("should export package stuff", () => {
 
     expect(index).toHaveProperty("ArrayValidator");
     expect(index).toHaveProperty("BooleanValidator");
+    expect(index).toHaveProperty("ConfigurationValidatorError");
+    expect(index).toHaveProperty("EnumerableValidator");
     expect(index).toHaveProperty("NumberValidator");
     expect(index).toHaveProperty("RegularExpressionValidator");
     expect(index).toHaveProperty("StringValidator");