+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2019-10-19
+
 ### Added
 
 - Handle object notation within top-level command line arguments and environment variables.
+- Enumerable validator.
 
 ### Changed